backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
python-oracledb==2.3.0
bcrypt==4.1.3
yfinance==0.2.40
pandas==2.2.2
python-dotenv==1.0.1
gunicorn==22.0.0
